Complete rework of cmake and remove visual studio solution
Cmake has been completely redone, with a more modern, readable and more organized code, allowing us to develop more easily.

Due to the fact that cmake has a modern and updated code from now on, I am removing the solution from visual studio, which is outdated and it is unnecessary to use a form of compilation that only works for windows if we can use cmake which is used for both windows and linux.