feat: add support for num_plays in animatedTexture (.aPNG) files (#398)
Addition of new functionality to the software by introducing support for the 'num_plays' variable in 'animatedTexture' (.aPNG) files. This variable determines the number of times an animation should loop, and was not previously supported by the software. These changes affect multiple files and include the addition of a new function, as well as modifications to existing functions, in order to fully implement the new feature. The new feature allows for greater flexibility in using animated textures, such as using them for effects on widgets in addition to just backgrounds.