improve: update GHA to read builtin-baseline from vcpkg.json (#403)
To prevent future build breaks due to changes in the vcpkg repository, added a step in the GitHub Actions to install vcpkg based on the builtin-baseline in the vcpkg.json file. This allows us to set vcpkg to be installed from specific commits in both GitHub Actions and in Ubuntu/Windows, centralizing everything in one location. We can only update vcpkg when there is an official release.