fix: added missing CMake module for asio library (#430)
When Boost got removed in favor of standalone Asio, it stopped working on CMake because Asio is not one of the modules usually shipped with CMake (Boost is), thus it failed to configure. Adds a simple script to look for Asio and link as expected. This problem in asio does not occur in windows or ubuntu, apparently only in some other systems, such as arch linux and mac OS.