Fix mouse buttons
As an example, MouseMidButton was the same as pressing MouseLeftButton+MouseRightButton at the same time.
Because MouseLeftButton (1) + MouseRightButton (2) = MouseMidButton (3). It means, if I check in my code if I'm pressing mid button `g_window.isPressed(MouseMidButton)`, it will work with mid button, but ALSO if I click left+right buttons.

It should not work like that.
Mouse buttons uses bitwise logic, so these values should be like 0, 1, 2, 4, 8, and so on, not like 0, 1, 2, 3, 4, 5, 6, 7, 8, and so on.