fix: correct search for keybind (eg. ctrl+e)
When searching for the keybind (eg.: ctrl+e) the search will not return results.

The find function in Lua is used to search for a pattern within a string. By default, find uses Lua's pattern matching, which is similar to regular expressions but with some differences. The + character in Lua patterns is a special character that matches one or more occurrences of the previous character.

We need to escape the character if it is present on the search text.