perf: Texture Atlas (#1252)
A Texture Atlas system was implemented on the client, resulting in a performance boost of 1.3x to 3x. For those not familiar with rendering, an atlas is essentially a large texture that combines multiple images used on screen, reducing texture switches and GPU communication overhead. The result? Stable FPS even with lots of objects and effects being rendered simultaneously.